alter security

+++ backend/app/controllers/plans.js
@@ -12,15 +12,22 @@ module.exports = function (app) {
      * @route /plans
      */
     controller.getPlans = async (req, res) => {
-        const userValid = await Auth.validaUser(req);
-        if (userValid) {
-            let plans;
-            if (userValid.level === 3) {
-                plans = await Plans.findAll();
+        try {
+            const userValid = await Auth.validaUser(req);
+            if (userValid) {
+                if (userValid.level !== 3) {
+                    return res.status(401).json({
+                        "message": "User cannot access this record"
+                    });
+                }
+                let plans = await Plans.findAll();
+                res.status(200).json(plans)
+            } else {
+                res.status(500).json("error: fail get plans")
             }
-            res.status(200).send(plans)
-        } else {
-            res.status(500).json("error: fail get plans")
+        } catch (error) {
+            console.log(error);
+            return res.status(500).json(error)
         }
     }
 
@@ -32,16 +39,26 @@ module.exports = function (app) {
 * @route /plans/:id
 */
     controller.getPlan = async (req, res) => {
-        const userValid = await Auth.validaUser(req);
-        if (userValid) {
-            let plans = await Plans.findAll({
-                where: {
-                    id: req.params.id
+        try {
+            const userValid = await Auth.validaUser(req);
+            if (userValid) {
+                if (userValid.level !== 3) {
+                    return res.status(401).json({
+                        "message": "User cannot access this record"
+                    });
                 }
-            });
-            res.status(200).send(plans[0])
-        } else {
-            res.status(500).json("error: fail get plans")
+                let plans = await Plans.findAll({
+                    where: {
+                        id: req.params.id
+                    }
+                });
+                res.status(200).json(plans[0])
+            } else {
+                res.status(500).json("error: fail get plans")
+            }
+        } catch (error) {
+            console.log(error);
+            return res.status(500).json(error)
         }
     }
 
@@ -53,42 +70,51 @@ module.exports = function (app) {
  * @route /plan
  */
     controller.addPlan = async (req, res) => {
-        const userValid = await Auth.validaUser(req);
-        if (userValid) {
-
-            let planExist = await Plans.findOne({
-                where: {
-                    name: req.body.name,
+        try {
+            const userValid = await Auth.validaUser(req);
+            if (userValid) {
+                if (userValid.level !== 3) {
+                    return res.status(401).json({
+                        "message": "User cannot access this record"
+                    });
                 }
-            });
-
-            if (planExist) {
-                let msg = `The plan are already registered!!! ID: ${planExist.dataValues.id}`
-                res.status(406).json(msg)
-
-            } else {
-                let save = await Plans.create({
-                    name: req.body.name,
-                    description: req.body.description,
-                    users_limit: req.body.users_limit,
-                    applications_limit: req.body.applications_limit,
-                    servers_limit: req.body.servers_limit
+                let planExist = await Plans.findOne({
+                    where: {
+                        name: req.body.name,
+                    }
                 });
-                if (save) {
-                    let values = []
-                    values.push({
-                        id: save.dataValues.id,
+
+                if (planExist) {
+                    let msg = `The plan are already registered!!! ID: ${planExist.dataValues.id}`
+                    res.status(406).json(msg)
+
+                } else {
+                    let save = await Plans.create({
                         name: req.body.name,
                         description: req.body.description,
                         users_limit: req.body.users_limit,
                         applications_limit: req.body.applications_limit,
                         servers_limit: req.body.servers_limit
                     });
-                    res.status(200).send(values);
+                    if (save) {
+                        let values = []
+                        values.push({
+                            id: save.dataValues.id,
+                            name: req.body.name,
+                            description: req.body.description,
+                            users_limit: req.body.users_limit,
+                            applications_limit: req.body.applications_limit,
+                            servers_limit: req.body.servers_limit
+                        });
+                        return res.status(200).json(values);
+                    }
                 }
+            } else {
+                return res.status(401).json("error: user invalid");
             }
-        } else {
-            res.status(401).send("error: user invalid");
+        } catch (error) {
+            console.log(error);
+            return res.status(500).json(error)
         }
     }
     /**
@@ -99,81 +125,101 @@ module.exports = function (app) {
      * @route /plans/:id
      */
     controller.updatePlan = async (req, res) => {
-        const userValid = await Auth.validaUser(req);
-        let msg;
-        if (userValid) {
-            let planExist = await Plans.findOne({
-                where: {
-                    name: req.body.name
+        try {
+            const userValid = await Auth.validaUser(req);
+            let msg;
+            if (userValid) {
+                if (userValid.level !== 3) {
+                    return res.status(401).json({
+                        "message": "User cannot access this record"
+                    });
                 }
-            });
-            if (planExist && planExist.dataValues.id !== req.body.id) {
-                msg = `The plan are already registered!!! ID: ${planExist.dataValues.id}`
-                res.status(406).json(msg)
-
-            } else {
-                let save = await Plans.update({
-                    name: req.body.name,
-                    description: req.body.description,
-                    users_limit: req.body.users_limit,
-                    applications_limit: req.body.applications_limit,
-                    servers_limit: req.body.servers_limit
-                }, {
+                let planExist = await Plans.findOne({
                     where: {
-                        id: req.body.id
+                        name: req.body.name
                     }
                 });
-                if (save) {
-                    let values = []
-                    values.push({
-                        id: req.body.id,
+                if (planExist && planExist.dataValues.id !== req.body.id) {
+                    msg = `The plan are already registered!!! ID: ${planExist.dataValues.id}`
+                    return res.status(406).json(msg)
+
+                } else {
+                    let save = await Plans.update({
                         name: req.body.name,
                         description: req.body.description,
                         users_limit: req.body.users_limit,
                         applications_limit: req.body.applications_limit,
                         servers_limit: req.body.servers_limit
+                    }, {
+                        where: {
+                            id: req.body.id
+                        }
                     });
-                    res.status(200).send(values);
+                    if (save) {
+                        let values = []
+                        values.push({
+                            id: req.body.id,
+                            name: req.body.name,
+                            description: req.body.description,
+                            users_limit: req.body.users_limit,
+                            applications_limit: req.body.applications_limit,
+                            servers_limit: req.body.servers_limit
+                        });
+                        return res.status(200).json(values);
+                    }
                 }
+            } else {
+                return res.status(401).json("error: user invalid");
             }
-        } else {
-            res.status(401).send("error: user invalid");
+        } catch (error) {
+            console.log(error);
+            return res.status(500).json(error)
         }
     }
 
-     /**
-     * deletePlan
-     * @param {Object} req
-     * @param {Object} res
-     * @method DELETE
-     * @route /plans/:id
-     */
+    /**
+    * deletePlan
+    * @param {Object} req
+    * @param {Object} res
+    * @method DELETE
+    * @route /plans/:id
+    */
     controller.deletePlan = async (req, res) => {
-        const userValid = await Auth.validaUser(req);
-        if (userValid) {
-            let plan = await Plans.findOne({
-                where: {
-                    id: req.params.id
+        try {
+            const userValid = await Auth.validaUser(req);
+            if (userValid) {
+                if (userValid.level !== 3) {
+                    return res.status(401).json({
+                        "message": "User cannot access this record"
+                    });
                 }
-            });
-            if (plan) {
-                plan = plan.dataValues;
-                let appDelete = await Plans.destroy({
+                let plan = await Plans.findOne({
                     where: {
                         id: req.params.id
                     }
                 });
-                if (appDelete) {
-                    res.status(200).send(plan);
-                } else {
-                    res.status(500).send("error: it was not possible to delete the data.");
-                }
+                if (plan) {
+                    plan = plan.dataValues;
+                    let appDelete = await Plans.destroy({
+                        where: {
+                            id: req.params.id
+                        }
+                    });
+                    if (appDelete) {
+                        return res.status(200).json(plan);
+                    } else {
+                        return res.status(500).json("error: it was not possible to delete the data.");
+                    }
 
+                } else {
+                    return res.status(500).json("error: record does not exist");
+                }
             } else {
-                res.status(500).send("error: record does not exist");
+                return res.status(401).json("error: user invalid");
             }
-        } else {
-            res.status(401).send("error: user invalid");
+        } catch (error) {
+            console.log(error);
+            return res.status(500).json(error)
         }
     }
 
@@ -185,19 +231,29 @@ module.exports = function (app) {
      * @route /plans
      */
     controller.deletePlans = async (req, res) => {
-        const userValid = await Auth.validaUser(req);
-        if (userValid) {
-            let ids = req.body.id
-            for (let i = 0; i < ids.length; i++) {
-                await Plans.destroy({
-                    where: {
-                        id: ids[i]
-                    }
-                });
+        try {
+            const userValid = await Auth.validaUser(req);
+            if (userValid) {
+                if (userValid.level !== 3) {
+                    return res.status(401).json({
+                        "message": "User cannot access this record"
+                    });
+                }
+                let ids = req.body.id
+                for (let i = 0; i < ids.length; i++) {
+                    await Plans.destroy({
+                        where: {
+                            id: ids[i]
+                        }
+                    });
+                }
+                return res.status(200).json(ids);
+            } else {
+                return res.status(401).json("error: user invalid");
             }
-            res.status(200).send(ids);
-        } else {
-            res.status(401).send("error: user invalid");
+        } catch (error) {
+            console.log(error);
+            return res.status(500).json(error)
         }
     }
